Fix conftest
