+++ ckanext-unckan/conftest.py
@@ -1,0 +1,3 @@
+pytest_plugins = [
+    u'ckanext.unckan.tests.fixtures',
+]
