+++ docker/ckan/files/ckan.ini
@@ -140,7 +140,7 @@ ckan.cors.origin_allow_all = false
 ckan.cors.origin_whitelist =
 
 ## Plugins Settings ############################################################
-ckan.plugins = push_errors unckan image_view pdf_view datastore datatables_view datapusher_plus tracking api_tracking superset announcements dbquery selfinfo citeproc
+ckan.plugins = push_errors unckan image_view pdf_view datastore datatables_view datapusher_plus activity tracking api_tracking superset announcements dbquery selfinfo citeproc
 ckan.resource_proxy.timeout = 5
 
 ## Front-End Settings ##########################################################
